services/retriever_service/requirements.txt
python-dotenv
qdrant-client
redis[asyncio]
